requirements.txt
fastapi
uvicorn
medmnist
mlflow
boto3
mangum
torchvision
torch
Pillow